refactor: Reorganize server-related files and imports
