+++ packages/framework/src/servers/express.ts
@@ -1,9 +1,9 @@
 import { type VercelRequest, type VercelResponse } from '@vercel/node';
 import { type Request, type Response } from 'express';
 
-import { NovuRequestHandler, ServeHandlerOptions } from './handler';
-import { Either } from './types';
-import { type SupportedFrameworkName } from './types';
+import { NovuRequestHandler, ServeHandlerOptions } from '../handler';
+import { Either } from '../types';
+import { type SupportedFrameworkName } from '../types';
 
 export const frameworkName: SupportedFrameworkName = 'express';
