+++ packages/framework/src/servers/remix.ts
@@ -1,6 +1,6 @@
-import { NovuRequestHandler, type ServeHandlerOptions } from './handler';
-import { type SupportedFrameworkName } from './types';
-import { getResponse } from './utils';
+import { NovuRequestHandler, type ServeHandlerOptions } from '../handler';
+import { type SupportedFrameworkName } from '../types';
+import { getResponse } from '../utils';
 
 export const frameworkName: SupportedFrameworkName = 'remix';
