+++ packages/framework/src/types/index.ts
@@ -4,9 +4,9 @@ export * from './cron.types';
 export * from './discover.types';
 export * from './event.types';
 export * from './execution.types';
-export * from './framework.types';
 export * from './health-check.types';
 export * from './schema.types';
+export * from './server.types';
 export * from './skip.types';
 export * from './step.types';
 export * from './subscriber.types';
