+++ packages/framework/tsup.config.ts
@@ -4,7 +4,7 @@ import { type SupportedFrameworkName } from './src';
 const frameworks: SupportedFrameworkName[] = ['h3', 'express', 'next', 'nuxt', 'sveltekit', 'remix'];
 
 export default defineConfig({
-  entry: ['src/index.ts', ...frameworks.map((framework) => `src/${framework}.ts`)],
+  entry: ['src/index.ts', ...frameworks.map((framework) => `src/servers/${framework}.ts`)],
   sourcemap: false,
   clean: true,
   treeshake: true,
